+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - use transaction in `Contact::add_or_lookup()` #4059
 - Organize the connection pool as a stack rather than a queue to ensure that
   connection page cache is reused more often. #4065
+- Use transaction in `update_blocked_mailinglist_contacts`. #4058
 
 ### Fixes
 - Start SQL transactions with IMMEDIATE behaviour rather than default DEFERRED one. #4063